Move content to own directory
